+++ DWin3DLib/Math/Vec3i.cpp
@@ -6,7 +6,7 @@ Vec3i::Vec3i(int x, int y, int z)
 }
 
 Vec3i::Vec3i(const Vec3f &v3f)
-    : x(v3f.x + 0.5L), y(v3f.y + 0.5L), z(v3f.z + 0.5L)
+    : x((int)(v3f.x + 0.5L)), y((int)(v3f.y + 0.5L)), z((int)(v3f.z + 0.5L))
 {
 }
